AI/faster-whisper/requirements.txt
faster-whisper
ctranslate2
tqdm
ffmpeg-python